Refactoring of room Selection
Split from one RoomSelection file into a component and a Selectionpage view.
